pyproject.toml
[tool.poetry]
name = "google_cloud_platform"
version = "0.1.0"
description = ""
authors = ["Rodrigo Lazarini Gil <rodrigo.lazarini.gil@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
google-cloud-storage = "^1.32.0"
google-cloud-bigquery = "^2.3.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"